TASK1-Question 3
The program does not include an underscore on the ‘boat cost’ variable but it includes an underscore in the if statement. Also in the else statement the the print method has no semi colons.The less hen greater than symbol is facing the wrong way. I provided  a fix to the code